+++ src/main/java/de/uib/configed/ConfigedMain.java
@@ -59,6 +59,8 @@ import de.uib.utils.Utils;
 import de.uib.utils.logging.Logging;
 import de.uib.utils.swing.ButtonTabComponent;
 import de.uib.utils.table.gui.BooleanIconTableCellRenderer;
+import de.uib.utils.table.gui.DeviceTypeIconTableCellRenderer;
+import de.uib.utils.table.gui.PlatfromIconTableCellRenderer;
 import de.uib.utils.userprefs.UserPreferences;
 
 public class ConfigedMain {
@@ -297,7 +299,6 @@ public class ConfigedMain {
 		executor.runInParallel(() -> persistenceController.getGroupDataService().retrieveAllGroupsPD());
 		executor.runInParallel(() -> persistenceController.getGroupDataService().retrieveAllObject2GroupsPD());
 		executor.runInParallel(() -> persistenceController.getProductDataService().retrieveDepotProductPropertiesPD());
-		executor.runInParallel(() -> persistenceController.getHealthDataService().retrieveHostsWithHealthCheck());
 		executor.waitForCompletion();
 
 		ExtraFrameController.reloadDialogs();
@@ -423,11 +424,8 @@ public class ConfigedMain {
 
 		if (selectedClients.size() == 1) {
 			mainFrame.getClientConfiguration().getClientInfoPanel().setClientID(selectedClients.get(0));
-			mainFrame.getClientConfiguration().getClientInfoPanel()
-					.updateHealthCheckActiveCheckBoxStatus(selectedClients.get(0));
 		} else {
 			mainFrame.getClientConfiguration().getClientInfoPanel().setClientID("");
-			mainFrame.getClientConfiguration().getClientInfoPanel().updateHealthCheckActiveCheckBoxStatus(null);
 		}
 
 		hostInfo.resetGui();
@@ -447,7 +445,7 @@ public class ConfigedMain {
 
 		Logging.info(this, "updateHostInfo, produce hostInfo  selectedClients.length ", selectedClients.size());
 
-		if (!selectedClients.isEmpty()) {
+		if (!selectedClients.isEmpty() && !pcinfos.isEmpty()) {
 			hostInfo.setValues(pcinfos.get(selectedClients.get(0)).getMap());
 
 			Logging.debug(this, "updateHostInfo, produce hostInfo first selClient ", selectedClients.get(0));
@@ -644,15 +642,13 @@ public class ConfigedMain {
 		Logging.info(this, "buildPclistTableModel host_displayFields ",
 				persistenceController.getHostDataService().getHostDisplayFields());
 
-		Set<Object> hostsWithActiveHealthCheck = persistenceController.getHealthDataService()
-				.getHostsWithActiveHealthCheck();
 		for (String clientId : clientIds) {
 			HostInfo pcinfo = pcinfos.get(clientId);
 			if (pcinfo == null) {
 				pcinfo = new HostInfo();
 			}
 
-			Map<String, Object> rowmap = pcinfo.getDisplayRowMap0();
+			Map<String, Object> rowmap = pcinfo.getDisplayRowMap();
 
 			String sessionValue = "";
 			if (sessionInfo.get(clientId) != null) {
@@ -661,7 +657,6 @@ public class ConfigedMain {
 
 			rowmap.put(HostInfo.CLIENT_SESSION_INFO_DISPLAY_FIELD_LABEL, sessionValue);
 			rowmap.put(HostInfo.CLIENT_CONNECTED_DISPLAY_FIELD_LABEL, isHostConnected(clientId));
-			rowmap.put(HostInfo.HEALTH_CHECK_ACTIVE_FIELD_LABEL, hostsWithActiveHealthCheck.contains(clientId));
 
 			List<Object> rowItems = new ArrayList<>();
 
@@ -759,11 +754,15 @@ public class ConfigedMain {
 				Icons.getIntellijIcon("checkmark", null), null);
 		BooleanIconTableCellRenderer opsiCheckMarkCellRenderer = new BooleanIconTableCellRenderer(
 				Icons.getIntellijIcon("checkmark", Globals.OPSI_OK), null);
+		PlatfromIconTableCellRenderer platformIconTableCellRenderer = new PlatfromIconTableCellRenderer();
+		DeviceTypeIconTableCellRenderer deviceTypeIconTableCellRenderer = new DeviceTypeIconTableCellRenderer();
 
 		configureColumn(HostInfo.CLIENT_CONNECTED_DISPLAY_FIELD_LABEL, opsiCheckMarkCellRenderer);
 		configureColumn(HostInfo.CLIENT_WAN_CONFIG_DISPLAY_FIELD_LABEL, defaultCheckMarkCellRenderer);
 		configureColumn(HostInfo.CLIENT_INSTALL_BY_SHUTDOWN_DISPLAY_FIELD_LABEL, defaultCheckMarkCellRenderer);
-		configureColumn(HostInfo.HEALTH_CHECK_ACTIVE_FIELD_LABEL, defaultCheckMarkCellRenderer);
+		configureColumn(HostInfo.CLIENT_HEALTH_CHECK_ACTIVE_DISPLAY_FIELD_LABEL, defaultCheckMarkCellRenderer);
+		configureColumn(HostInfo.CLIENT_OS_TYPE_DISPLAY_FIELD_LABEL, platformIconTableCellRenderer);
+		configureColumn(HostInfo.CLIENT_DEVICE_TYPE_DISPLAY_FIELD_LABEL, deviceTypeIconTableCellRenderer);
 	}
 
 	private void configureColumn(String fieldLabel, TableCellRenderer renderer) {