+++ src/main/java/de/uib/opsidatamodel/serverdata/dataservice/ConfigDataService.java
@@ -198,11 +198,6 @@ public class ConfigDataService {
 		return cacheManager.getCachedData(CacheIdentifier.WAN_CONFIGURATION, Map.class);
 	}
 
-	public Map<String, List<Object>> getNotWanConfigurationPD() {
-		retrieveWANConfigOptionsPD();
-		return cacheManager.getCachedData(CacheIdentifier.NOT_WAN_CONFIGURATION, Map.class);
-	}
-
 	public Map<String, ConfigOption> retrieveWANConfigOptionsPD() {
 		Map<String, ConfigOption> allWanConfigOptions = extractSubConfigOptionsByInitial(
 				OpsiServiceNOMPersistenceController.CONFIG_KEY + "." + WAN_PARTKEY);
@@ -212,7 +207,6 @@ public class ConfigDataService {
 		Map<String, ConfigOption> notWanConfigOptions = extractSubConfigOptionsByInitial(
 				OpsiServiceNOMPersistenceController.CONFIG_KEY + "." + NOT_WAN_CONFIGURED_PARTKEY + ".");
 
-		Map<String, List<Object>> notWanConfiguration = new HashMap<>();
 		Map<String, List<Object>> wanConfiguration = new HashMap<>();
 
 		List<Object> values = null;
@@ -220,14 +214,9 @@ public class ConfigDataService {
 		for (Entry<String, ConfigOption> notWanConfigOption : notWanConfigOptions.entrySet()) {
 			if (notWanConfigOption.getValue().getType() != ConfigOption.TYPE.BOOL_CONFIG) {
 				Logging.error(this, "WAN config option key ", notWanConfigOption.getKey(), " is non BOOL_CONFIG");
-				notWanConfiguration.put(notWanConfigOption.getKey(), null);
 				wanConfiguration.put(notWanConfigOption.getKey(), null);
 			} else {
 				Boolean b = (Boolean) notWanConfigOption.getValue().getDefaultValues().get(0);
-
-				values = new ArrayList<>();
-				values.add(b);
-				notWanConfiguration.put(notWanConfigOption.getKey(), values);
 
 				values = new ArrayList<>();
 				values.add(!b);
@@ -237,8 +226,6 @@ public class ConfigDataService {
 
 		cacheManager.setCachedData(CacheIdentifier.WAN_CONFIGURATION, wanConfiguration);
 		Logging.info(this, "getWANConfigOptions wanConfiguration ", wanConfiguration);
-		cacheManager.setCachedData(CacheIdentifier.NOT_WAN_CONFIGURATION, notWanConfiguration);
-		Logging.info(this, "getWANConfigOptions notWanConfiguration  ", notWanConfiguration);
 
 		return allWanConfigOptions;
 	}
@@ -296,84 +283,6 @@ public class ConfigDataService {
 		persistenceController.notifyPanelCompleteWinProducts();
 	}
 
-	public List<Map<String, Object>> addWANConfigStates(String clientId, List<Map<String, Object>> jsonObjects) {
-		return addWANConfigStates(clientId, true, jsonObjects);
-	}
-
-	private List<Map<String, Object>> addWANConfigStates(String clientId, boolean wan,
-			List<Map<String, Object>> jsonObjects) {
-		retrieveWANConfigOptionsPD();
-
-		Map<String, List<Object>> wanConfiguration = getWanConfigurationPD();
-		Map<String, List<Object>> notWanConfiguration = getNotWanConfigurationPD();
-
-		Logging.debug(this, "addWANConfigState  wanConfiguration ", wanConfiguration, "\n ", wanConfiguration.size());
-		Logging.debug(this, "addWANConfigState  wanConfiguration.keySet() ", wanConfiguration.keySet());
-
-		Logging.debug(this, "addWANConfigState  notWanConfiguration ", notWanConfiguration, "\n ",
-				notWanConfiguration.size());
-		Logging.debug(this, "addWANConfigState  notWanConfiguration.keySet() ", notWanConfiguration.keySet());
-
-		setConfig(notWanConfiguration);
-		Logging.info(this, "set notWanConfiguration members where no entry exists");
-		// send to opsiserver only new configs
-		setConfig(true);
-
-		Map<String, List<Object>> specifiedConfiguration;
-
-		if (wan) {
-			specifiedConfiguration = wanConfiguration;
-		} else {
-			specifiedConfiguration = notWanConfiguration;
-		}
-
-		if (jsonObjects == null) {
-			jsonObjects = new ArrayList<>();
-		}
-
-		for (Entry<String, List<Object>> config : specifiedConfiguration.entrySet()) {
-			Logging.info(this, "addWANConfigState configId ", config.getKey());
-			Map<String, Object> item = Utils.createNOMitem(OpsiServiceNOMPersistenceController.CONFIG_STATE_TYPE);
-
-			item.put(OpsiServiceNOMPersistenceController.CONFIG_ID, config.getKey());
-
-			Logging.info(this, "addWANConfigState values ", config.getValue());
-
-			item.put(OpsiServiceNOMPersistenceController.VALUES_ID, config.getValue());
-
-			item.put(OpsiServiceNOMPersistenceController.OBJECT_ID, clientId);
-
-			Logging.info(this, "addWANConfigState configId, item ", config.getKey(), ", ", item);
-
-			// locally, hopefully the RPC call will work
-			if (getHostConfigsPD().get(clientId) == null) {
-				Logging.info(this, "addWANConfigState; until now, no config(State) existed for client ", clientId,
-						" no local update");
-				getHostConfigsPD().put(clientId, new HashMap<>());
-			}
-
-			getHostConfigsPD().get(clientId).put(config.getKey(), config.getValue());
-
-			// prepare for JSON RPC
-			jsonObjects.add(item);
-		}
-
-		return jsonObjects;
-	}
-
-	public boolean setWANConfigs(String clientId, boolean wan) {
-		boolean result = false;
-		Logging.info(this, "setWANConfigs ", clientId, " . ", wan);
-
-		List<Map<String, Object>> jsonObjects = addWANConfigStates(clientId, wan, null);
-
-		OpsiMethodCall omc = new OpsiMethodCall(RPCMethodName.CONFIG_STATE_UPDATE_OBJECTS,
-				new Object[] { jsonObjects });
-		result = exec.doCall(omc);
-
-		return result;
-	}
-
 	// send config updates and clear the collection
 	public void updateConfigs() {
 		setConfig(false);
@@ -426,7 +335,8 @@ public class ConfigDataService {
 		for (Map<String, Object> callConfig : configCollection) {
 			if (callConfig.get("defaultValues") == null) {
 				callsConfigDeleteCollection.add(callConfig);
-			} else if (!restrictToMissing || usedConfigIds.contains(callConfig.get("ident"))) {
+				callsConfigUpdateCollection.removeIf(item -> callConfig.get("ident").equals(item.get("ident")));
+			} else if ((!restrictToMissing || usedConfigIds.contains(callConfig.get("ident")))) {
 				callConfig.put("defaultValues", callConfig.get("defaultValues"));
 				callConfig.put("possibleValues", callConfig.get("possibleValues"));
 				callsConfigUpdateCollection.add(callConfig);
@@ -875,50 +785,6 @@ public class ConfigDataService {
 		return findBooleanConfigurationComparingToDefaults(host, wanConfiguration);
 	}
 
-	/**
-	 * Checks if the given clients have an entry for UEFI boot. That means, the
-	 * client has a client config with an entry
-	 * {@code clientconfig.uefinetbootlabel}.
-	 *
-	 * @param clients for which to check the existence of UEFI boot entry
-	 * @return null if clients have different values or if client list is empty
-	 */
-	@SuppressWarnings({ "java:S2447" })
-	public Boolean isUEFI43(Iterable<String> clients) {
-		Boolean isUEFI = null;
-
-		for (String client : clients) {
-			Map<String, Object> clientConfig = getHostConfigsPD().get(client);
-			if (clientConfig == null) {
-				isUEFI = false;
-				continue;
-			}
-
-			Object uefiConfig = clientConfig.get("clientconfig.uefinetbootlabel");
-
-			if (uefiConfig instanceof List && !((List<?>) uefiConfig).isEmpty()) {
-				if (Boolean.FALSE.equals(isUEFI)) {
-					return null;
-				} else {
-					isUEFI = true;
-				}
-			} else {
-				if (Boolean.TRUE.equals(isUEFI)) {
-					return null;
-				} else {
-					isUEFI = false;
-				}
-			}
-		}
-
-		return isUEFI;
-	}
-
-	public boolean configureInstallByShutdown(String clientId, boolean shutdownInstall) {
-		return setHostBooleanConfigValue(OpsiServiceNOMPersistenceController.KEY_CLIENTCONFIG_INSTALL_BY_SHUTDOWN,
-				clientId, shutdownInstall);
-	}
-
 	// for checking if WAN default configuration is set
 	public boolean findBooleanConfigurationComparingToDefaults(String host,
 			Map<String, List<Object>> defaultConfiguration) {
@@ -962,26 +828,6 @@ public class ConfigDataService {
 			Logging.debug(this, "valueFromConfigStateAsExpected ", result);
 		}
 		return result;
-	}
-
-	private boolean setHostBooleanConfigValue(String configId, String hostName, boolean val) {
-		Logging.info(this, "setHostBooleanConfigValue ", hostName, " configId ", configId, " val ", val);
-
-		List<Object> values = new ArrayList<>();
-		values.add(val);
-
-		Map<String, Object> item = Utils.createNOMitem(OpsiServiceNOMPersistenceController.CONFIG_STATE_TYPE);
-		item.put(OpsiServiceNOMPersistenceController.OBJECT_ID, hostName);
-		item.put(OpsiServiceNOMPersistenceController.VALUES_ID, values);
-		item.put(OpsiServiceNOMPersistenceController.CONFIG_ID, configId);
-
-		List<Map<String, Object>> jsonObjects = new ArrayList<>();
-		jsonObjects.add(item);
-
-		OpsiMethodCall omc = new OpsiMethodCall(RPCMethodName.CONFIG_STATE_UPDATE_OBJECTS,
-				new Object[] { jsonObjects });
-
-		return exec.doCall(omc);
 	}
 
 	public Boolean getGlobalBooleanConfigValue(String key, Boolean defaultVal) {